Fixes #19
--README.md
Modified Retrieve Organisation Data Endpoint section to show an example of how to obtain product description, and markup its text when embedded on the web page. Additional set HTML escaping to show how to avoid HTML injection attacks.

--APIv1Util.php
Created class to store common utility functions that can be used for the library

--APIv1ExampleRunnerRetrieveOrgESDData.php
Modified example class to show how to obtain product description, and markup its text when embedded on the web page. Additionally set HTML escaping to show how to avoid HTML code injection attacks.

--APIv1ExampleRunnerRetrieveOrgESDDataCategories.php
Modified example class to show how to markup category descriptions when embedded on the web page to display with the styling set in the data.

--APIv1ExampleRunnerRetrieveOrgESDDataProductImages.php
Modified example class to show how to markup product image descriptions when embedded on the web page to display with the styling set in the data.

--APIv1ExampleRunnerRetrieveOrgESDDataProductProducts.php
Modified example class to show how to markup product description1 when embedded on the web page to display with the styling set in the data.